Add Config#supervise & disable supervision in CLI, Thor, & Rake tasks.
When a CLI action is executed (or the test action in the Thor and Rake
tasks), there is no need to supervise the instance actors and restart
them after crashes. Actually leaving the actors crashed is exactly the
behavior we want.

However, supervision (restarting crashed instance actors) is desirable
in the console or when the codebase is being used as a third party
library. Therefore supervision will remain turned on by default, but
can be disabled by setting `config.supervise = false` before calling the
instances factory/accessor method.