Synchronize librarian-chef/berks install commands.
When both of these tools are run initially with no .lock file present
they are going to generate this file. It would be dangerous having
multiple copies of Berkshelf or Librarian active at the same time, so
we'll allow one thread to execute at a time with Jamie's shared mutex.