Refactor Kitchen::Kb to be a concern of Driver::Base only, not Instance.
While doing the renaming refactor of Jr -> Kb it surprised me how
classes where at least partially involved with Kb's business. Since this
is a helper object with minimal state intended for drivers it makes
sense for Driver::Base to own it.

I couldn't see an easy way to constructor-inject this object dependency
as Kb needs to know the suite's name. So for the moment Driver::Base will
create this object itself.