Install Omnibus if require_chef_omnibus version differs from current.
This will allow a user to upgrade or downgrade their Omnibus package
version to match the version string in require_chef_omnibus. If the
version string matches what is currently installed, the installation is
skipped.

**Note** This commit also supports a value of `"latest"` for
require_chef_omnibus which drops the version flag when calling the omnibus
installation script. Since calculating the lastest Chef release is
nontrivial, the Omnibus installer will be called on every converge when
`"latest"` is set--whether Chef is installed or not.

Closes #32