Add support for nodes.
Assuming a directory of `test/integration/nodes/` containing JSON files,
these will be added to the provisioner's sandbox for transfer to the
instance.

Note that this behavior is identical to data bags and roles in that a
suite-specific `nodes/` directory will be used if it exists. For
example, given a suite name of `client` the directory
`test/integration/client/nodes/` will be used for the provisioner's
sandbox.