Initial Chef Zero provisioner: use chef-client on your instances!
This is a first-pass implementation of a chef-client provisioning mode
using Chef Zero, a tiny API-complete implementation of a Chef Server.
The strategy this provisioner uses is the following:

1. Install Chef Omnibus package
2. Create a local sandbox of all Chef data, same as chef_solo except for
   the addition of a client.rb and the chef-client-zero.rb shim binary
3. Install the chef-zero and knife-essentials gems into the Omnibus Ruby
4. Run the chef-client-zero.rb binary which provides all normal
   configuration availble to chef-client

The chef-client-zero.rb shim starts up a Chef Zero server instance
running in the background, overrides some basic configuration (such as
chef_server_url, client_key, and validation_key), creates a chef-client
application and runs it. At the end of the chef-client run the Chef Zero
server will be stopped.