ensure Busser and Chef-Zero execute out of sandboxes
This change ensures the main Omnibus Chef install is not polluted
with gems that are only leveraged for provisioning or test execution.
There are also situations where we cannot install gems into the
Omnibus Chef install location as it may be read-only (see Chef for
Junos).

This commit includes the following enhancements:

* Install all gems into sandboxes.
* Ensure all remotely executed commands are sandbox aware. This is
accomplished by setting GEM_HOME and GEM_PATH for all remote commands.