Ensure chef_zero provisioner respects require_chef_omnibus version.
The current strategy here is to install a modern version of Chef into
the gem sandbox directory outside of the Omnibus installation in order
to get Chef Zero, the Knife Essentials code which was folded into Chef.

Next we change the chef-client-zero.rb wrapper script to setup the Chef
Zero server and finally fork-and-exec the `chef-client` run in order to
use the `chef-client` binary found in PATH. This would normally
correspond to /usr/bin/chef-client which is a symlink to the Omnibus
Chef package install in most cases. Otherwise if Chef is installed
somewhere else found on the path, this will be used instead.

Granted installing Chef twice over (once for the Omnibus install and
once by gem install for Chef Zero) is tedious but this will be remedied
in a later commit to only use the chef-client-zero.rb wrapper if the
desired version of Chef is less than 11.8.0.

Closes #243
References #240
References #224