Use `chef-client -z` for Chef versions >= 11.8.0 in chef_zero mode.
Older versions of Chef will fall back to the wrapper script strategy.
The version of Chef is determined by inspecting the value of the
`require_chef_omnibus` config option. All values are assumed to result
in a version of Chef greater than or equal to 11.8.0 unless an explicit
version string is given that is less than 11.8.0. In other words, values
of `true`, `nil`, and `"latest"` will all assume a modern Chef version
capable of supporting local mode in chef-client (a.k.a. chef-zero mode).

Closes #224