Test Kitchen now works out of `/tmp/kitchen` for all providers.
Previous to this commit, Test Kitchen's work directory on the instance
depended on the provisioner used. For example the chef_solo provisioner
worked out of `/tmp/kitchen-chef-solo` and the chef_zero provisioner
worked out of `/tmp/kitchen-chef-zero`.

With the addition of a `data_path` config option this made discovering
the path to uploaded data more difficult. Now all provisioners will
upload the contents of `data_path` to `/tmp/kitchen/data`.