DataMunger - introduce kitchen config block select options.
Currently there are two overridable values that can go in a `kitchen`
block: `kitchen_root` and `test_base_path`. A `kitchen` block only
exists in the root of the data structure and is subject to overriding
by the initialized instance of Kitchen::Config.

For example, the following is supported:

    ---
    kitchen:
      # default is Dir.pwd
      kitchen_root: /some/other/root
      # default is "#{Dir.pwd}/test/integration"
      test_base_path: acceptance

These configuration values will be populated in every Provisioner,
Driver, and Busser configuration. These will be set by DataMunger
and cannot be overriden like other configuration.