Support older version of Chef with chef_zero Provisioner, woo!
To make this work, a modern version of Chef will be installed into a
sandboxed path (no dependencies at all on Omnibus packages gems) and
used to start chef-zero on a thread during a normal chef-client run
(using your desired Chef version).

This is best-effort for versions of Chef less than 11.8.0, there may
be missing package dependencies needed on your specific platform.
If this is the case, the chef_zero Provisioner may crash and fail
the converge action.

Note that versions of Chef greater than or equal to 11.8.0 will be
supported out of the box with Test Kitchen running chef-client in
local mode to get the desired functionality.