Generate a more explict form for driver & provisioner in `kitchen init`.
There is a simple/default for for specify the driver and provisioner
names like so:

    driver: vagrant
    provisioner: chef_zero

which is really equivalent to:

    driver:
      name: vagrant
    provisioner:
      name: chef_zero

Since the latter is easier to add to and doesn't require any knowledge,
it should probably be the default form that newcomers see for the first
time.

Thanks to @webframp for the discussion.