Expose Provisioner#sandbox_path for use by drivers & remove tmpdir.
This commit changes the behavior of `Provisioner#create_sandbox` in that
it no longer returns the created sandbox path. Now the expected calling
behavior is to call `Provisioner#create_sandbox` and then refer to the
sandbox path with `Provisioner#sandbox_path`. This allows
`#create_sandbox` to make use of calls to super in a clean way and to
decouple the command and query nature of this method.