Add local & global file locations with environment variables.
Previous to this commit you could override the location of your
`.kitchen.yml` with the `KITCHEN_YAML` environment variable. For
example:

    KITCHEN_YAML=Kitchen.yaml kitchen list

Now the local and global config file locations can also be overridden
with the following environment variables:

* `KITCHEN_LOCAL_YAML`
* `KITCHEN_GLOBAL_YAML`

The local override might be useful in a hosted CI environment where your
configuration overrides need to be added into your project repository.
You could therefore setup an environment variable such as:

    KITCHEN_LOCAL_YAML=.kitchen.ci.yml

That wouldn't be used in day-to-day local development.

Thanks to @coderanger and @damm for the idea and use case.

Closes #304