Default to use Chef formatter output for solo & zero provisioners.
The current default behavior of chef-client & chef-solo runs is to use
the default output formatter and a logger with level "auto" (note, not
"info"). Thus, when running Kitchen with its default logging level
("info"), the resulting Chef runs will use log level "auto". When
running Kitchen with a log level other than "info", this value will be
passed down to the `--log_level` flag for Chef. In this way we still get
debug level output when required.