Disable color output when no TTY is present.
This should help clean up kitchen output in CI environments, using
the kitchen command with unix pipes, etc.

In other words:

    kitchen test | cat

Should *not* output colors.

Currently, the only catch is that colors are still used inside the
instances for tools such as RSpec. This may require work in the Busser
component.

Closes #330