Remove TravisCI/JRuby testing until support is more explicit.
Support for running Test Kitchen on JRuby should actually be pretty
close, but needs a bit of time to ensure nothing has been missed. Seeing
as how this Ruby version always fails and slows down the test matrix,
we'll drop this for the time being in favor of a quicker feedback cycle.

Then...one day soon...oh yes...