Populate Chef Zero provisioner cookbook cache
The Chef Zero provisioner can run the test client in local
mode, but when the test client is first created and invoked
it still needs to download cookbooks from the Chef Zero
server. I have found this initial "Synchronizing Cookbooks"
to be slow when there are many cookbooks in the test.

This change copies the cookbooks used to populate the Chef
Zero server into the kitchen cache directory, which is the
cache directory of the test client. When the client runs
its initial "Synchronizing Cookbooks", the cookbooks will
already be in the cache and this step will be faster.