Backfill Transport support for Driver::SSHBase.
This commit aims to provide backwards compatibility for the SSHBase
Driver superclass upon which most existing drivers are based. The
`Kitchen::SSH` object is swapped out for a newly-wired Transport which
for the moment is assumed to be `Transport::Ssh`. No existing methods
are removed, no existing method signatures are modified. There are 3
remaining methods which are meant to be used with `Kitchen::SSH`:

* `#run_remote` uses a `Kitchen::SSH` connection and rescues the
  relevant exceptions
* `#transfer_path` also uses a `Kitchen::SSH` connection and rescues the
  relevant exceptions
* `#build_ssh_args` returns an Array structure used to populate the
  `Kitchen::SSH` constructor

These above methods are preserved as-is for Driver subclasses which may
call them.