[Potentially Breaking] Provisioners responsible for converge action.
Potentially Breaking Notes
==========================

This is potentially breaking to Driver authors if all of the following
are true:

* Your Driver currently directly inherits from `Kitchen::Driver::Base`
* Your Driver implements/overrides the `#converge` method

Put another way, your Driver may have issues if it looks like the
following:

    module Kitchen
      module Driver
        class MyDriver < Kitchen::Driver::Base
          def conerge(state)
            # custom converge work
          end
        end
      end
    end

For the vast majority (well over 90%) of OSS Drivers in the wild,
current behavior is maintained as they all inherit from
`Kitchen::Driver::SSHBase`. This class has been cemented to preserve its
current behavior, and Test Kitchen will invoke the `#converge` method
for these Drivers.

**Note:** upgrade path and instructions for Driver authors will be
written, but backwards compatibility is being taken seriously.

A future deprecation process may remove the `SSHBase` backwards
compatibility, but not without plenty of lead time and warning. Due to
the constraints of SemVer, by definition, this wouldn't occur before a
2.x codebase release.

Self-Aware Provisioners
=======================

As of this commit a `#call(state)` method exists in
`Kitchen::Provisioner::Base` which will be invoked by Test Kitchen when
the converge action is performed. For backwards compatibility, the same
convergence "template" is used, relying on a small number of public
methods that return command strings and 3 methods responsible for
sandbox creation and cleanup.

The high-level description of the default `#call(state)` method is as
follows:

1. Create the temporary sandbox on the workstation with "stuff" to
   transfer to the remote instance.
2. Run the `#install_command` on the remote instance, if it is
   implemented.
3. Run the `#init_command` on the remote instance, if it is
   implemented.
4. Transfer all files in the sandbox path to the remote instance's
   configured `:root_path`.
5. Run the `#prepare_command` on the remote instance, if it is
   implemented.
6. Run the `#run_command` on the remote instance, if it is
   implemented.

As a Provisioner author, you may elect to overwrite or partially
re-implement the `#call(state)` method to do whatever you need in
whatever order makes sense. This key difference allows Provisioner
authors to entirely own the `kitchen converge` action and not also rely
on the Driver used to manage the instances.