[Transport::Winrm] #execute now only raises when exit code is != 0.
Before this commit, #execute would be considered a failure if the
command had an exit code of 0 but had data in the stderr stream. Using
the Transport day-to-day lead to the conclusion that there were too many
false positives which the previous interpretation and developers
authoring PowerShell code must take care to handle exit codes correctly.