New plugin type: Verifier with Busser being the first implementation.
Verifiers are a type of plugin like Drivers, Provisioners, and
Transports. They recieve configuration, can calculate default values,
and be loaded from external RubyGems.

For backwards compatibility, Busser will be the first Verifier and
Busser will be the default Verifier. Any `:busser` configuration block
will be converted into a `:verifer` block.

More work is still required to finish the new abstraction, but the
system passes tests, is bootable, and works for newer and legacy
Drivers.