[Potentially Breaking] Verifiers responsible for verify action.
Potentially Breaking Notes
==========================

This is a potentially breaking change to Driver authors if all of the
following are true:

* Your Driver currently directly inherits from `Kitchen::Driver::Base`
* Your Driver implements/overrides the `#setup` and/or `#verify` methods

Put another way, your Driver may have issues if it looks like the
following:

    module Kitchen
      module Driver
        class MyDriver < Kitchen::Driver::Base
          def setup(state)
            # custom setup work
          end

          def verify(state)
            # custom verify work
          end
        end
      end
    end

For the vast majority (well over 90%) of OSS Drivers in the wild,
current behavior is maintained as they all inheit from
`Kitchen::Driver::SSHBase`. This class has been cemented to preserve its
current behavior, and Test Kitchen will invoke the `#setup` and
`#verify` methods for these Drivers.

**Note:** upgrade path and instructions for Driver authors will be
written, but backwards compatibility is being taken seriously.

A future deprecation process may remote the `SSHBase` backwards
compatibility, but not without plenty of lead time and warning. Due to
the constraints of SemVer, by definition, this wouldn't occur before a
2.x codebase release.

Self-Aware Verifiers
====================

As of this commit a `#call(state)` method exists in
`Kitchen::Verifier::Base` which will be invoked by Test Kitchen when the
verify action is performed. The setup action which previously installed
the Busser gem and plugins, becomes a dummy or "no-op" action. In other
words all previous behavior in the setup action now takes place in the
verify action. For backwards compatibility, the same verify "template"
is used, relying on a small number of public methods that return strings
and 3 new methods responsible for sandbox creation and cleanup (with
very similar implemention to that in Provisioners).

The high-level description of the default `#call(state)` method is as
follows:

1. Create the temporary sandbox on the workstation with "stuff" to
   tranfer to the remote instance.
2. Run the `#install_command` on the remote instance, if it is
   implemented.
3. Run the `#init_command` on the remote instance, if it is
   implemented.
4. Transfer all files in the sandbox path to the Verifier's configured
   `:root_path` on the remote instance.
5. Run the `#prepare_command` on the remote instance, if it is
   implemented.
6. Run the `#run_command` on the remote instance, if it is
   implemented.

As a Verifier author, you may elect to overwrite or partially
re-implement the `#call(state)` method to do whatever you need in
whatever order makes sense. This key difference allows Verifier authors
to entirely own the `kitchen verify` action and also not rely on the
Driver used to manage the instances.