[ChefSolo,ChefZero] Reload PATH on Windows for older Omnibus packages.
This is to deal with older Omnibus packages, only on Windows. The batch
file wrappers for `chef-solo.bat` and `chef-client.bat` don't fully
qualify the path to `ruby.exe` and instead rely on `PATH`. Note: later
versions of Omnibus packages fixed this.

Now that Test Kitchen reuses a remote shell session, the PATH is not
updated after the Omnibus package is installed. This commit reloads
`PATH` into the PowerShell script which invokes `chef-solo.bat` and
`chef-client.bat` commands.