[Busser] Prefix explicit path to ruby.exe in install_command.
This would make all calls to `gem` work like the following for
PowerShell:

    & \path\to\bin\ruby.exe \path\to\bin\gem install busser

This way we can avoid any windows stub scripts (.bat files) that do not
have a fully qualified path to `ruby.exe`.

An issue was raised when using older versions of Chef Omnibus packages
(such as 11.16.x) which contain older versions of RubyGems or packages
which rely too much on `ruby.exe` being set in `PATH`.