Add :sudo_command to Provisioners, Verifiers, & ShellOut.
In an effort to better support older distros (such as CentOS 5), and
other distros which don't ship with sudo in $PATH (such as Solaris),
a new configuration attribute is introduced into `Provisioner::Base` and
`Verifier::Base` called `:sudo_command`. For greatest portability and
backwards compatibility this defaults to `"sudo -E"` but is now
customizable depending on your situation.

For example:

    ---
    driver:
      name: vagrant

    platforms:
      - name: centos-7.0                      # defaults apply
      - name: centos-5.11                     # removes -E flag
        provisioner:
          sudo_command: sudo
        verifier:
          sudo_command: sudo
      - name: solaris-10                      # sets custom path to sudo
        provisioner:
          sudo_command: /usr/local/bin/sudo
        verifier:
          sudo_command: /usr/local/bin/sudo

Note that a future feature proposes a way to remove the
provisioner/verifier duplication but this would only be a convenience,
not a behavior change to this commit.

Finally, `Driver::Base` was not augemented with this configuration
attribute as it is no longer responsible for creating the commands to
execute on remote instances--this is now soley the purview of
Provisioner and Verifier plugins.

Closes #592
Closes #629
Closes #307
References #321