[CommandExecutor] Move ObjectSpace finalizer logic into executor.
Any remote shell cleanup needs to be handled by the CommandExecutor so
that it attempts to close the last opened shell. Previous to this
commit, the shell cleanup was in `Kitchen::Transport::Winrm` but only
knew about the **initially** opened shell.

The `ShellCloser` class was extracted to full class status as the user
of `CommandExecutor` now has to supply a closer in the constructor if
ObjectSpace cleanup behavior is desired (the default is to do no
cleanup).

Closes #656