[Transport::Winrm] Extract dependant code to winrm-transport gem.
This commit extracts the implementation logic for the WinRM Transport
and all of its associated direct and transitive gem dependencies into
the 'winrm-transport' gem. Prior to this commit, Test Kitchen had a
direct dependency on the following gems:

* winrm
* rubyzip

Now, all gem dependencies for the `CommandExecutor` and
`FileTransporter` are managed in the winrm-transport gem. This code (and
its requisite dependencies) are **only** required if the Test Kitchen
project is using the Winrm Transport, and so will only be activated in
this case. All other projects that do not require WinRM will remain
unaffected and Test Kitchen can retain its minmal gem dependency graph
which includes a highly desirable goal of no native gems in its graph.

If the winrm-transport gem is not present, `kitchen` commmands will
crash with a UserError message instructing the user to install the gem
or add it to the project's Gemfile. For ChefDK users, a future release
will include the winrm-transport gem and so should not be affected by
this in the long term.