Adds number_of_password_prompts as a SSH option
This allows this option of Net::SSH to be changed from within a driver. This
is needed when a VM template spins up an ssh daemon when the cloud-provided
password isn't set yet. This would prompt for a password, while this should
just fail.

The cloudstack driver needs this option to allow wait_for_sshd to wait for
an actual working sshd, and the possibility to login with the user/pass combo.

I assume other drivers might benefit from this too.