Ensure sandbox_path is not within kitchen_root.
In rare cases the system temp directory becomes unusable, which results
in the mktmpdir trying to use the current working directory. This
becomes an issue as the default value for kitchen_root is also cwd
which results in a confusing cyclical error in which it attempts to copy
the cookbooks into a temp directory within the path which is is
copying. This should avoid most instances of this error, and present a
possible solution in simple terms.

Resolves Issue #779