Fix for net-ssh 2.10.0.
This sets the default compression to auto-negotiate the best available.
This is broken currently because net-ssh no longer silently allows
falling back to no compression and OpenSSH uses `"zlib@openssh.com"`,
not `"zlib"`. I've removed the validation as it is incorrect and in the
future it would be nice to not need an ASAP release for this kind of
issue. If a user misconfigured the compression setting, they will now
notice immediately because it will refuse to connect.