Running the chef_base provisioner install_command via sudo, and adding support for a command prefix for provisioners and verifiers
Currently, whether or not to use sudo during the installation of
the omnibus chef package is left up to the install command
generated by Mixlib::Install based on the sudo flag and command
passed to the constructor. Future versions of Mixlib::Install
do not support the sudo option, and some platforms (such as Cisco
Nexus) require the entire process to run with escalated privileges
in order to obtain outbound internet access. This change tells
Mixlib::Install to not care about sudo and instead wrap the entire
install_command with sudo from within the provisioner.

On certain platforms, commands must be prefixed with another
command to work properly, such as on Cisco Nexus where the
user must switch into a management vrf to obtain outbound network
access. Using the sudo_command config parameter is not sufficient
as some commands, such as busser verifier commands and provisioner
init commands, need to run as the non-privileged user.

This change allows the user to configure a string that will be
prepended to provisioner and verifier commands, much like how
sudo_command works, but impacts all commands.