Upgrade to Aruba 0.11.1/Cucumber 2.1.0 without warnings.
This should repair the damaged features suite when a newer version of
Aruba was used. There were deprecation warnings, but worse a change in
the `run_simple`/environment loading, leading to executed commands which
were not getting the benefit of a properly prepared `ENV` hash.

Thanks to @iennae for the pairing session to finally get this sorted!

Closes #790