Adding a username method to the SSH and WinRM transport classes
In some cases, an instance's driver might need knowledge of the
transport username in order to perform a task, such as resetting
the password for the user on a GCE Windows instance.

This change exposes the username via a public method to avoid
such hacks as accessing the `#config` private method.