support passing Kitchen::Config Hash keys to Kitchen::RakeTasks.new
The current implementation of the `Kitchen::RakeTasks.new` method is that you aren't able to modify the config generated by `Kitchen::Config.new`. As such, if you want to change the `test_base_path` value, you need to duplicate most of this helper in to your `Rakefile`.

This change adds the ability to pass a Hash in to the `.new` method, and it will be merged with the config sent to `Kitchen::Config.new`. If nothing is passed in to the `.new` method, it defaults to an empty Hash (`{}`).