Driver cache interface for leverage in provisioner
Changing the implementation of the cache directory, now we have decided
to create a new interface in the Driver abstraction so that they can
implement it when they are ready to consume it.

The Provisioner now will wait for that cache interface to be implemented
so that it leverages it inside ChefBase, it will detect it and pass the
`-d` option to Omnibus.

We had to implement the cache interface inside SSHBase so that the tests
passes because Proxy is still inheriting from it and not the Base.

Signed-off-by: Salim Afiune <afiune@chef.io>