Fix Dir.glob usage
Dir.glob is broken on Windows in a variety of ways. Changing to the
directory you want to glob before doing any sort of globbing is the safest
way to use it from past experience.

Related to https://github.com/test-kitchen/test-kitchen/issues/1238 in
that it is the very next error one would run into with berkshelf being
fixed.

Signed-off-by: Jay Mundrawala <jmundrawala@chef.io>