include list of known plugins when one cannot be loaded
In an effort to be friendlier, let's include a list of plugins kitchen
can find when it cannot load the one named in a config.

Because the require assumes the plugin will be in the load path at
`kitchen/{plugin-type}`, that is what is used to walk the load path
looking for any Ruby file present at that location that defines a class.

Closes #879

Signed-off-by: Robb Kidd <robb@thekidds.org>