Adds kitchen-dokken template override when driver specified
To provide a better experience for the kitchen dokken users it would be
helpful if there was a way to generate a kitchen configuration with more
of the fields required to make it work. As this plugin is actually several
that act in concert this may be the right approach.

It also may not be the right approach. In the situation that someone wants
to specify multiple drivers. There were also some thoughts about the `init`
command going away completely.

Perhaps this template should instead be the the default of the Chef DK.

Let's discuss.

Signed-off-by: Franklin Webber <franklin@chef.io>