Add a pre_install_command to all provisioners.
This is useful for addressing any kind of bootstrapping problem that
might occur prior to provisioner installation that requires custom code.

For example, it is often useful to coordinate the provisioner
installation with actions that start at boot on an instance via some
kind of init system (e.g. systemd, cloud-init).