using an if as it is better for a multiple condition.
changing how local_config is found, in the past, it was based on the config file, and it still is to a point, but because there are 2 methods, we calculate the dot and undot, and then check if the dot exists and if does we return it, and if it doesn't we return the undot.  This is more or less the same form as the primary config.

Signed-off-by: Vern Burton <me@vernburton.com>